+++ src/json_codecs.ts
@@ -3,5 +3,6 @@ export {Converter} from 'caesium-core/converter';
 
 export {JsonObject,JsonQuery} from './json_codecs/interfaces';
 export {str, num, bool, date, list, map} from './json_codecs/basic';
+export {enumToString} from './json_codecs/enum_to_string';
 export {model} from './json_codecs/model_to_json';
 export {union} from './json_codecs/model_union_to_json';