+++ test/src/json_codecs.spec.ts
@@ -1,4 +1,5 @@
 import {basicTests} from './json_codecs/basic.spec';
+import {enumToStringTests} from './json_codecs/enum_to_string.spec';
 import {modelPropertyToJsonTests} from './json_codecs/model_property_to_json.spec';
 import {modelToJsonTests} from './json_codecs/model_to_json.spec';
 import {modelUnionToJsonTests} from './json_codecs/model_union_to_json.spec';
@@ -8,6 +9,7 @@ import {objectToJsonTests} from "./json_codecs/object_to_json.spec";
 export function jsonCodecsTests() {
     describe('json_codecs', () => {
         basicTests();
+        enumToStringTests();
         objectToJsonTests();
         modelPropertyToJsonTests();
         modelToJsonTests();