+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1
 - JsonObject, JsonQuery, Codec, Converter, identity now exported from json_codecs module
 - onData, thisArg removed from ResponseHandler and thisArg added to BaseResponseHandler
+- Fixed ModelManager resolution error
 
 ## 0.1.0
 Breaking changes