+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
-## 0.1.1
+## 0.2.0
+- Added `enumToString` codec to the json_codecs module
+
+### Bugfixes
 - JsonObject, JsonQuery, Codec, Converter, identity now exported from json_codecs module
 - onData, thisArg removed from ResponseHandler and thisArg added to BaseResponseHandler
 - Fixed ModelManager resolution error
+
 
 ## 0.1.0
 Breaking changes